@@ -5,21 +5,21 @@ import 'package:flutter/material.dart';
 import 'utils.dart';
 
 class BasePainter extends CustomPainter {
-  Color baseColor;
-  Color selectionColor;
-  int primarySectors;
-  int secondarySectors;
-  double sliderStrokeWidth;
+  late Color baseColor;
+  late Color selectionColor;
+  late int primarySectors;
+  late int secondarySectors;
+  late double sliderStrokeWidth;
 
-  Offset center;
-  double radius;
+  late Offset center;
+  late double radius;
 
   BasePainter({
-    @required this.baseColor,
-    @required this.selectionColor,
-    @required this.primarySectors,
-    @required this.secondarySectors,
-    @required this.sliderStrokeWidth,
+    required this.baseColor,
+    required this.selectionColor,
+    required this.primarySectors,
+    required this.secondarySectors,
+    required this.sliderStrokeWidth,
   });
 
   @override
@@ -63,7 +63,7 @@ class BasePainter extends CustomPainter {
     }
   }
 
-  Paint _getPaint({@required Color color, double width, PaintingStyle style}) =>
+  Paint _getPaint({required Color color, double? width, PaintingStyle? style}) =>
       Paint()
         ..color = color
         ..strokeCap = StrokeCap.round
