@@ -12,42 +12,42 @@ enum SlidingState { none, endIsBiggerThanStart, endIsSmallerThanStart }
 typedef SelectionChanged<T> = void Function(T a, T b, T c);
 
 class CircularSliderPaint extends StatefulWidget {
-  final CircularSliderMode mode;
-  final int init;
-  final int end;
-  final int divisions;
-  final int primarySectors;
-  final int secondarySectors;
-  final SelectionChanged<int> onSelectionChange;
-  final SelectionChanged<int> onSelectionEnd;
-  final Color baseColor;
-  final Color selectionColor;
-  final Color handlerColor;
-  final double handlerOutterRadius;
-  final Widget child;
-  final bool showRoundedCapInSelection;
-  final bool showHandlerOutter;
-  final double sliderStrokeWidth;
-  final bool shouldCountLaps;
+  late final CircularSliderMode mode;
+  late final int init;
+  late final int end;
+  late final int divisions;
+  late final int primarySectors;
+  late final int secondarySectors;
+  late final SelectionChanged<int> onSelectionChange;
+  late final SelectionChanged<int> onSelectionEnd;
+  late final Color baseColor;
+  late final Color selectionColor;
+  late final Color handlerColor;
+  late final double handlerOutterRadius;
+  late final Widget child;
+  late final bool showRoundedCapInSelection;
+  late final bool showHandlerOutter;
+  late final double sliderStrokeWidth;
+  late final bool shouldCountLaps;
 
   CircularSliderPaint({
-    @required this.mode,
-    @required this.divisions,
-    @required this.init,
-    @required this.end,
-    this.child,
-    @required this.primarySectors,
-    @required this.secondarySectors,
-    @required this.onSelectionChange,
-    @required this.onSelectionEnd,
-    @required this.baseColor,
-    @required this.selectionColor,
-    @required this.handlerColor,
-    @required this.handlerOutterRadius,
-    @required this.showRoundedCapInSelection,
-    @required this.showHandlerOutter,
-    @required this.sliderStrokeWidth,
-    @required this.shouldCountLaps,
+    required this.mode,
+    required this.divisions,
+    required this.init,
+    required this.end,
+    required this.child,
+    required this.primarySectors,
+    required this.secondarySectors,
+    required this.onSelectionChange,
+    required this.onSelectionEnd,
+    required this.baseColor,
+    required this.selectionColor,
+    required this.handlerColor,
+    required this.handlerOutterRadius,
+    required this.showRoundedCapInSelection,
+    required this.showHandlerOutter,
+    required this.sliderStrokeWidth,
+    required this.shouldCountLaps,
   });
 
   @override
@@ -58,21 +58,21 @@ class _CircularSliderState extends State<CircularSliderPaint> {
   bool _isInitHandlerSelected = false;
   bool _isEndHandlerSelected = false;
 
-  SliderPainter _painter;
+  late SliderPainter _painter;
 
   /// start angle in radians where we need to locate the init handler
-  double _startAngle;
+  double _startAngle = 0.0;
 
   /// end angle in radians where we need to locate the end handler
-  double _endAngle;
+  double _endAngle = 0.0;
 
   /// the absolute angle in radians representing the selection
-  double _sweepAngle;
+  double _sweepAngle = 0.0;
 
   /// in case we have a double slider and we want to move the whole selection by clicking in the slider
   /// this will capture the position in the selection relative to the initial handler
   /// that way we will be able to keep the selection constant when moving
-  int _differenceFromInitPoint;
+  late int _differenceFromInitPoint;
 
   /// will store the number of full laps (2pi radians) as part of the selection
   int _laps = 0;
@@ -244,7 +244,7 @@ class _CircularSliderState extends State<CircularSliderPaint> {
   }
 
   void _handlePan(Offset details, bool isPanEnd) {
-    RenderBox renderBox = context.findRenderObject();
+    RenderBox renderBox = context.findRenderObject() as RenderBox;
     var position = renderBox.globalToLocal(details);
 
     var angle = coordinatesToRadians(_painter.center, position);
@@ -283,7 +283,7 @@ class _CircularSliderState extends State<CircularSliderPaint> {
     if (_painter == null) {
       return false;
     }
-    RenderBox renderBox = context.findRenderObject();
+    RenderBox renderBox = context.findRenderObject() as RenderBox;
     var position = renderBox.globalToLocal(details);
 
     if (position == null) {
@@ -332,9 +332,9 @@ class CustomPanGestureRecognizer extends OneSequenceGestureRecognizer {
   final Function onPanEnd;
 
   CustomPanGestureRecognizer({
-    @required this.onPanDown,
-    @required this.onPanUpdate,
-    @required this.onPanEnd,
+    required this.onPanDown,
+    required this.onPanUpdate,
+    required this.onPanEnd,
   });
 
   @override
